pyproject.toml
[project]
name = "mqtt-palyground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "django-extension>=0.1.0",
    "psycopg>=3.0.0",
    "python-dotenv>=1.0.0",
    "paho-mqtt>=2.0.0",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
]
